Pass runRuleLength and maxNoOfExclusions explicitly in results functions.
NB I have left noRegrets and overhangingReversions at default, because passing them explicitly in results_functions could be confusing given that other algorithm control arguments are passed through in get_results. And it is reasonably clear from the documentation that with noRecals = FALSE and noPeriodMin = FALSE, you are getting SSA with no regrets and overhanging reversions.